LPS-86000 Move into exception package
