+++ modules/apps/login/login-authentication-google-web/src/main/java/com/liferay/login/authentication/google/web/internal/portlet/action/GoogleLoginErrorMVCRenderCommand.java
@@ -27,7 +27,7 @@ import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.security.sso.google.GoogleAuthorization;
-import com.liferay.portal.security.sso.google.StrangersNotAllowedException;
+import com.liferay.portal.security.sso.google.exception.StrangersNotAllowedException;
 
 import javax.portlet.PortletException;
 import javax.portlet.RenderRequest;
