+++ modules/apps/portal-security-sso/portal-security-sso-google-api/bnd.bnd
@@ -4,4 +4,5 @@ Bundle-Version: 3.0.0
 Export-Package:\
 	com.liferay.portal.security.sso.google,\
 	com.liferay.portal.security.sso.google.configuration,\
-	com.liferay.portal.security.sso.google.constants
+	com.liferay.portal.security.sso.google.constants,\
+	com.liferay.portal.security.sso.google.exception
